requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.12.3

# Tools
django-extensions==2.2.8

# Testing
mypy==0.761
pytest==5.3.5
pytest-sugar==0.9.2
pytest-django==3.8.0
factory-boy==2.12.0

# Code quality
flake8==3.7.9
